+++ digi/xbee/recovery.py
@@ -1,4 +1,4 @@
-# Copyright 2019-2021, Digi International Inc.
+# Copyright 2019-2023, Digi International Inc.
 #
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -32,7 +32,11 @@ SUPPORTED_HARDWARE_VERSIONS = (HardwareVersion.XBEE3.code,
                                HardwareVersion.XBEE3_SMT.code,
                                HardwareVersion.XBEE3_TH.code,
                                HardwareVersion.XBEE3_RR.code,
-                               HardwareVersion.XBEE3_RR_TH.code)
+                               HardwareVersion.XBEE3_RR_TH.code,
+                               HardwareVersion.XBEE3_DM_LR.code,
+                               HardwareVersion.XBEE3_DM_LR_868.code,
+                               HardwareVersion.XBEE_XR_900_TH.code,
+                               HardwareVersion.XBEE_XR_868_TH.code)
 
 _BAUDRATE_KEY = "baudrate"
 _PARITY_KEY = "parity"